Deprecate listmarket in favour of listmarkets
