+++ cmd/tdex/listmarket.go
@@ -1,9 +1,7 @@
 package main
 
 import (
-	"context"
-
-	pboperator "github.com/tdex-network/tdex-daemon/api-spec/protobuf/gen/operator"
+	"fmt"
 
 	"github.com/urfave/cli/v2"
 )
@@ -15,20 +13,8 @@ var listmarket = cli.Command{
 }
 
 func listMarketAction(ctx *cli.Context) error {
-	client, cleanup, err := getOperatorClient(ctx)
-	if err != nil {
-		return err
-	}
-	defer cleanup()
-
-	resp, err := client.ListMarket(
-		context.Background(), &pboperator.ListMarketRequest{},
+	return fmt.Errorf(
+		"this command is deprecated and will be removed in the next version.\n" +
+			"Instead, use the new command 'tdex listmarkets'",
 	)
-	if err != nil {
-		return err
-	}
-
-	printRespJSON(resp)
-
-	return nil
 }
